cmd: adding files and stages, commit uses index
The add command is able to add new files to a stage (depricating
"track") and add the new stage to the index/commit list, or add
existing stages to the index/commit list. The getRoot function is
used to get any parent directories with a .duc folder, which is used
to locate the config and index. The commit command now accepts no args
and will commit every stage in the commitSet. After commiting, the
commitSet in the index is cleared.

closes #4
closes #5