commit: checksum and commit dir manifests as JSON
This commit makes directoryManifest structs get committed to the cache
using the same subroutine as regular files. Prior to this change,
directoryManifests were first checksummed using encoding/gob and then
saved in the cache as JSON. In other words, if you were to checksum the
JSON-encoded directoryManifest file in the cache, it would not match the
checksum as addressed in its file path.  This discrepancy with other
cached files is strange and unexpected, and it would unnecessarily
complicate a future cache validation routine.