checkout, commit: continue recursive dir support
Choosing whether or not to recurse directories is decided by the
Artifact, not the operation (e.g. commit, checkout); this creates
a cleaner user experience: The user would `duc add -r my_dir/` and then
commit, checkout, etc. as usual; the knowledge of which artifacts are
recursive is tracked by DUC, not the user.