add ArtifactWithStatus
Previously, the status of an Artifact was a completely separate struct
from the Artifact itself. When beginning to implement artifacts which do
not use the DUC cache, it became clear that the final interpretation of
the Status struct requires knowledge of the Artifact struct itself --
notably the IsDir and new SkipCache boolean fields.