add basic integration/end-to-end testing framework
The hope with this testing framework is to provide a better/faster way to
develop and debug high-level tests. The current integration tests are
overly complex and difficult to scale (e.g. to directory
commit/checkout/status). This framework aims solve these problems by
allowing the developer to write tests in shell commands, emulating
direct user interaction with the filesystem and the duc executable.
I could see this approach becoming instrumental in translating bug
reports to test cases.