begin Index overhaul, tweak Stage.Run method
The Index now owns all tracked Stages in its in-memory data structure,
and it loads said stages when it itself is loaded from the filesystem.
With the Index owning all Stages, we have a means to look up what
Artifacts are owned by what Stages, and thus we can begin implementing
the DAG of Stages for various important capabilities (e.g. recursive
Stage.Run).