cd to project root before running top-level cmds
The integration tests caught a bug where directory manifests were being
committed with absolute paths rather than paths relative to the project
root. This bug was introduced when rootDir was added to all Stage
methods, as rootDir was always an absolute path. The sensible fix was to
always chdir to the project root and let the underlying code assume as
much.