dirManifest.Contents now a map, rm old dir tests
Making directoryManifest.Contents a map whose keys are Artifact paths
simplifies the directory handling code significantly. This refactor
spurred the removal of the bloated directory tests using mocks, as the
directoryManifest was tightly coupled to the tests. It is a TODO to add
more non-mocked tests around directory functionality, especially status
and checkout.