first pass at Index.Status()
It's looking like Index will supplant Stage as the main interface for
the CLI frontend. Incidentally, the Stage tests (which weren't very
helpful anyways) will be replaced by tests of the Index. Initially
I started to mock out Stages in the Index tests, but it quickly became
clear that Stages and the Index are tightly coupled, and mocking would
do more harm than good. I settled on removing the parrot Stage tests in
favor of better tests in the Index layer.