remove Index.ToCommit, refocus cmd/add
The commit set is no longer used to select which Stages to commit, and
as it is not particularly missed, I'm removing it entirely (for now).
Consequently, I've also made cmd/add only responsible for adding Stages
to the Index. I'm hoping a more general Stage creation CLI can be
created in the future (see #28).

closes #25