make Stage.FromFile more robust
I realized that wholesale using or rejecting a lock file in
Stage.FromFile was incorrect. This behavior led to a bug where a simple
modification to a Stage file (e.g. adding another output) would
essentially discard all Artifact checksums -- even those that were
currently checked out. The fix was to do away with Stage.IsEquivalent
and perform the equivalence check on each Artifact individually. This
work also uncovered a subtle bug with deserialization which is fixed and
commented in Stage.FromFile.

Also in this commit:
* switched to kabob-case for YAML fields
* began rewriting the README