add --pin, switch to tree in integration tests
The new --pin flag in run_tests.py allows you to quickly overwrite all
expected results with the current behavior in a test. This is meant to
simplify the workflow of updating tests with benign differences (e.g.
a slight change in Stage deserialization). It obviously should not
replace manual inspection of test output state.

Additionally, the tree utility replaces the homegrown find command used
to generate a summary of the filesystem.