read dir contents lazily instead of using ioutil
In the many_small_files benchmark, this method has comparable runtime to
the non-lazy ioutil.ReadDir method. I didn't profile memory usage
(definitely an important TODO), but it's certainly dangerous not to cap
the number of FileInfo structs in memory at a given time (just as it's
dangerous not to cap the number of worker goroutines). This is reason
enough to pull this code in. We should profile this code before
optimizing much further.