rename project to Dud
The name "duc" is already used by another open-source project:
https://github.com/zevv/duc . The project is well-established and
already owns the Ubuntu package `duc`. To avoid confusion, I'm renaming
the project Dud (pronounced "duhd", not "dood"). I chose this name
because it is reasonably close to the original name, it remains
easy to say (still one syllable), and it pays homage to Git with its
playful irreverence.