improve worker scheduling for dir commits
This commit changes from scheduling a fixed number of workers per
directory to dynamically scheduling workers for a directory and all
sub-directories. The fixed worker approach ran both the risk of
scheduling many more workers than were needed (particularly in the case
of many small nested directories) and the risk of scheduling less
workers than would be optimal (in the case of very large directories).
This solution improves on that model by limiting the total number of
workers available to a top-level artifact and all of its children and
by scheduling workers on-demand rather than upfront.