robustify check to move files during commit
This commit fixes a rare bug where committing fails because the cache
and workspace exist on the same underlying device, but on different
mount points; thus os.Rename returns an EXDEV errno. The fix is to
abandon the previous approach (check the files' device IDs) and simply
try moving a canary file between the workspace and the cache.